requirements.txt
# Web Framework
Flask==3.0.0
gunicorn==21.2.0

# Scraping
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0

# Template Engine
Jinja2==3.1.3

# Scheduling
APScheduler==3.10.4

# Image Processing
Pillow==10.2.0

# Utilities
python-dotenv==1.0.0
pytz==2024.1

# Testing
pytest==8.0.0
pytest-cov==4.1.0
responses==0.25.0
